requirements.txt
langchain
langchain-openai
langchain-community
gradio
python-dotenv
sqlalchemy
langsmith